update names processing

+++ 02_scrape_names.py
@@ -1,7 +1,5 @@
 # %%
 import os
-import time
-import pandas as pd
 from scrape_wikidata.names import (
     SQL_GN_SAID_TO_BE_SAME_AS,
     SQL_FN_SAID_TO_BE_SAME_AS,
@@ -16,6 +14,7 @@ from scrape_wikidata.names import (
 
 
 from pathlib import Path
+
 out_folder = "out_data/wikidata/raw/persons/by_dob"
 
 
@@ -29,12 +28,13 @@ Path(BASE_PATH_FN).mkdir(parents=True, exist_ok=True)
 
 # %%
 
-
+# Scrape first name said to be the same as
 for page in range(0, 100, 1):
     pagesize = 5000
     path = os.path.join(
         BASE_PATH_GN,
-        f"stbtsa_page_{page}_{page*pagesize}_to_{(page+1)*pagesize-1}.parquet")
+        f"stbtsa_page_{page}_{page*pagesize}_to_{(page+1)*pagesize-1}.parquet",
+    )
 
     if not os.path.exists(path):
         df = get_standardised_table(
@@ -90,39 +90,43 @@ for page in range(0, 100, 1):
             path,
             index=False,
         )
-    time.sleep(20)
 
 # %%
 
 for page in range(0, 100, 1):
     pagesize = 5000
-    df = get_standardised_table(SQL_GN_SHORTNAME, "shortname", page, pagesize)
-    df = df.drop_duplicates()
     path = os.path.join(
         BASE_PATH_GN,
         f"shortname_page_{page}_{page*pagesize}_to_{(page+1)*pagesize-1}.parquet",
     )
-    df.to_parquet(
-        path,
-        index=False,
-    )
-    time.sleep(20)
+
+    if not os.path.exists(path):
+        df = get_standardised_table(SQL_GN_SHORTNAME, "shortname", page, pagesize)
+        df = df.drop_duplicates()
+
+        df.to_parquet(
+            path,
+            index=False,
+        )
 
 
 # %%
 for page in range(0, 100, 1):
     pagesize = 5000
-    df = get_standardised_table(SQL_HYPOCORISM, "hypocorism", page, pagesize)
-    df = df.drop_duplicates()
     path = os.path.join(
         BASE_PATH_GN,
         f"hypo_page_{page}_{page*pagesize}_to_{(page+1)*pagesize-1}.parquet",
     )
-    df.to_parquet(
-        path,
-        index=False,
-    )
-    time.sleep(20)
+    if not os.path.exists(path):
+
+        df = get_standardised_table(SQL_HYPOCORISM, "hypocorism", page, pagesize)
+        df = df.drop_duplicates()
+
+        df.to_parquet(
+            path,
+            index=False,
+        )
+
 # %%
 diminutives = get_diminutives()
 path = os.path.join(
@@ -131,27 +135,5 @@ path = os.path.join(
 )
 diminutives.to_parquet(
     path,
-    index=False,
-)
-
-# %%
-
-names = pd.read_parquet("scrape_wikidata/raw_data/names/name_type=given/")
-counts = pd.read_parquet(
-    "scrape_wikidata/processed_data/step_6_births_namefreq/name_counts_from_births_register.parquet"
-)
-df = get_given_name_weighted_lookup(names, counts)
-df.to_parquet(
-    "scrape_wikidata/processed_data/step_3_alt_name_lookups/given_name.parquet",
-    index=False,
-)
-
-family_names = pd.read_parquet("scrape_wikidata/raw_data/names/name_type=family/")
-family_counts = pd.read_parquet(
-    "scrape_wikidata/processed_data/step_6_births_namefreq/surname_counts_from_births_register.parquet"
-)
-df = get_family_name_weighted_lookup(family_names, family_counts)
-df.to_parquet(
-    "scrape_wikidata/processed_data/step_3_alt_name_lookups/family_name.parquet",
     index=False,
 )
